AccountManagerPlugin: Change sequence of actions when deleting a user, refs #8685.
Nathaniel McCallum noticed, that with `SessionStore` db storage backend
the 'user deleted' event never fires due to wrong ordering of actions.
The call to `store.delete_user` is more elegant - pythonic - now, but a bit
more subtle to me too.

Thanks for spotting this and even more for providing a fix as well.

git-svn-id: file:///home/jun66j5/var/svn/trac-hacks.org/accountmanagerplugin/trunk@11250 7322e99d-02ea-0310-aa39-e9a107903beb