AccountManagerPlugin: Hotfix for pending user-registration issue, refs #9079, #9252, #9843 and #9090.
Now the reversion of [10520] done by [10585] is completed, although there are
many related changes and improvements in-between, so this is not quite the
same code as before.

Taking the chance to update now obsolete configuration examples in README too.

This has already open for much too long, so let's ditch the rather esoteric
considerations of `authenticated` user entries in Trac db table `session`
for now and move on.

Big sorry, that this took that much time, and thanks to Peter Stuge for
finally pushing me to it tonight.

git-svn-id: file:///home/jun66j5/var/svn/trac-hacks.org/accountmanagerplugin/trunk@11826 7322e99d-02ea-0310-aa39-e9a107903beb