AccountManagerPlugin: Extend `IAccountRegistrationInspector` interface including fallback, refs #8930.
While its fine to by-pass the rather new Python doc-string extraction, that
requires a recent Trac, the 'doc' attribute is still new and not guaranteed
to be present in all `IAccountRegistrationInspector` implementations.

This has been reported specifically by slav0nic for
`tracspamfilter.accountadapter.RegistrationFilterAdapter`, the pioneer
implementation of that interface outside !AccountManagerPlugin.
So it seems sensible to do both, consequently extend the interface definition
and implement a fallback for the early adopter.

Added some more unit tests, to ensure that the fallback works as expected.
While at it, one of the `ExtensionOrder` methods go a more intuitive name.

git-svn-id: file:///home/jun66j5/var/svn/trac-hacks.org/accountmanagerplugin/trunk@12663 7322e99d-02ea-0310-aa39-e9a107903beb