AccountManagerPlugin: Change class doc-string extraction for i18n to actually work.
This especially reverts and reworks parts of [12589] and [12663].
The exploration after cboos's hint in t:#11088 made me turn round and use an
approach closely resembling what already had been done to translate docs for
Trac core WikiMacros, although I changed it to fit for Trac 0.12 as well.

Trac core functions are used by default, but the configuration is tweaked
intentionally to ''not'' create catalogs in the '*-js' and 'tracini' domain.
So everything is still collected in 'messages', even if three compiler runs
on the same set of catalogs and generating unneeded message javascripts look
a bit awkward.

Because compatibility code predominated in `acct_mgr.util` it was about time
to re-arrange and form a dedicated `acct_mgr.compat` module.

git-svn-id: file:///home/jun66j5/var/svn/trac-hacks.org/accountmanagerplugin/trunk@12721 7322e99d-02ea-0310-aa39-e9a107903beb